sanitize messages too
